% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-utils.R
\name{argument_type_to_r}
\alias{argument_type_to_r}
\title{Returns the R type of an argument in the declarations}
\usage{
argument_type_to_r(argument)
}
\arguments{
\item{argument}{an argument like `declarations()[[1]]$arguments[[1]]`}
}
\description{
Returns the R type of an argument in the declarations
}
\note{
Returns `NA` when the argument doesn't have an R equivalent
}
