% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp-utils.R
\name{cpp_return_statement}
\alias{cpp_return_statement}
\title{Creates a return statement for a torch cpp function.}
\usage{
cpp_return_statement(returns)
}
\arguments{
\item{returns}{the returns object as in `declarations()[[1]]$returns`}
}
\description{
Creates a return statement for a torch cpp function.
}
