% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp-utils.R
\name{cpp_use_argument}
\alias{cpp_use_argument}
\title{Returns how the argument should be used in the cpp code.}
\usage{
cpp_use_argument(argument)
}
\arguments{
\item{argument}{an argument element as in `declarations()[[1]]$arguments[[1]]`}
}
\description{
For example: to use `Rcpp::XPtr<torch::Tensor> x` one must
refer to the pointer.
}
\details{
It will return `NA` if we don't know how to deal with this argument type.
The user must know how to deal with the `NA`s.
}
