% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{declarations}
\alias{declarations}
\title{Reads and get's the declaration file.}
\usage{
declarations()
}
\description{
Uses the options `torchgen.version` or `torchgen.path` to get a
Declarations.yaml file. If the `torchgen.path` is specified it's
always used instead of the `version`.
}
