% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{declarations_with_name}
\alias{declarations_with_name}
\title{Filters the declarations getting only those with
the same name.}
\usage{
declarations_with_name(declarations, name)
}
\arguments{
\item{declarations}{list of methods/function contained
in the declaration.}

\item{name}{name to filter.}
}
\description{
Filters the declarations getting only those with
the same name.
}
