% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{function_names}
\alias{function_names}
\title{Takes an object like `declarations()` and
returns the list of unique names in it.}
\usage{
function_names(declarations)
}
\arguments{
\item{declarations}{list of methods/function contained
in the declaration.}
}
\description{
Takes an object like `declarations()` and
returns the list of unique names in it.
}
