% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s4.R
\name{generic_s4_code}
\alias{generic_s4_code}
\title{Generates the S4 generic definition}
\usage{
generic_s4_code(methods)
}
\arguments{
\item{methods}{the methods list. something like
`tensor_methods() %>% declaration_with_name("abs")`}
}
\description{
Generates the S4 generic definition
}
