% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6-utils.R
\name{get_default_values}
\alias{get_default_values}
\title{Returns a named list with default values for all arguments.}
\usage{
get_default_values(methods)
}
\description{
Returns a named list with default values for all arguments.
}
