% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-utils.R
\name{get_possible_argument_names}
\alias{get_possible_argument_names}
\title{Get possible arguments for all overloaded methods.
The returned order is **important**!}
\usage{
get_possible_argument_names(methods)
}
\arguments{
\item{methods}{the methods list. something like
`tensor_methods() %>% declaration_with_name("abs")`}
}
\description{
Get possible arguments for all overloaded methods.
The returned order is **important**!
}
