% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cpp.R
\name{method_cpp_body}
\alias{method_cpp_body}
\title{Makes the body of the cpp function.}
\usage{
method_cpp_body(method)
}
\arguments{
\item{method}{a single element of the [declarations()] list.}
}
\description{
Makes the body of the cpp function.
}
