% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cpp.R
\name{method_cpp_code}
\alias{method_cpp_code}
\title{Creates the cpp code for a method.}
\usage{
method_cpp_code(method)
}
\arguments{
\item{method}{a single element of the [declarations()] list.}
}
\description{
Creates the cpp code for a method.
}
