% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cpp.R
\name{method_cpp_create_out}
\alias{method_cpp_create_out}
\title{Returns the code to create an `out` variable using a method and it's
arguments.}
\usage{
method_cpp_create_out(method)
}
\description{
Returns the code to create an `out` variable using a method and it's
arguments.
}
