% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cpp.R
\name{method_cpp_name}
\alias{method_cpp_name}
\title{Makes the name of the cpp function.}
\usage{
method_cpp_name(method)
}
\arguments{
\item{method}{a single element of the [declarations()] list.}
}
\description{
Makes the name of the cpp function.
}
