% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cpp.R
\name{method_cpp_return_type}
\alias{method_cpp_return_type}
\title{Gets the return type of the cpp function.}
\usage{
method_cpp_return_type(method)
}
\arguments{
\item{method}{a single element of the [declarations()] list.}
}
\description{
Gets the return type of the cpp function.
}
