% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cpp.R
\name{method_cpp_signature}
\alias{method_cpp_signature}
\title{Makes the signature of the cpp function.}
\usage{
method_cpp_signature(method)
}
\arguments{
\item{method}{a single element of the [declarations()] list.}
}
\description{
Makes the signature of the cpp function.
}
