% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_r6.R
\name{method_r6_code}
\alias{method_r6_code}
\title{Generate wraper code given a collection of methods with
the same name.}
\usage{
method_r6_code(methods)
}
\arguments{
\item{methods}{A collection of methods as you would get with
`tensor_methods() %>% declarations_with_name("abs")`}
}
\description{
Generate wraper code given a collection of methods with
the same name.
}
