% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s4.R
\name{method_s4_argument_names_and_types}
\alias{method_s4_argument_names_and_types}
\title{Returns a tibble of argument names and types of methods.
Already discarding argument names with `NA` that wound't be
converted.}
\usage{
method_s4_argument_names_and_types(method)
}
\description{
Returns a tibble of argument names and types of methods.
Already discarding argument names with `NA` that wound't be
converted.
}
