% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s4.R
\name{method_s4_signature}
\alias{method_s4_signature}
\title{Creates the signature as required by S4 methods in R
See [methods::setMethod()]'s `signature` argument.}
\usage{
method_s4_signature(method, methods = tensor_methods())
}
\arguments{
\item{method}{a single element of the [declarations()] list.}
}
\description{
Creates the signature as required by S4 methods in R
See [methods::setMethod()]'s `signature` argument.
}
