% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_s4.R
\name{namespace_s4_code}
\alias{namespace_s4_code}
\title{Generates the definition of an S4 method}
\usage{
namespace_s4_code(method)
}
\arguments{
\item{method}{a single element of the [declarations()] list.}
}
\description{
Generates the definition of an S4 method
}
