% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_s4.R
\name{namespace_s4_generic_name}
\alias{namespace_s4_generic_name}
\title{Get's the name of the generic function corresponding to the method}
\usage{
namespace_s4_generic_name(method)
}
\arguments{
\item{method}{a single method or a list of methods.}
}
\description{
Get's the name of the generic function corresponding to the method
}
