% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s4.R
\name{method_s4_generic_name}
\alias{method_s4_generic_name}
\title{Get's the name of the generic function corresponding to the method}
\usage{
method_s4_generic_name(method)
}
\arguments{
\item{method}{a single method or a list of methods.}
}
\description{
Get's the name of the generic function corresponding to the method
}
